move stk to media-libs, add examples useflag, don't install project headers and .cpp files

+++ overlays/proaudio/media-libs/stk/stk-4.3.1.ebuild
@@ -11,13 +11,20 @@ SRC_URI="http://ccrma.stanford.edu/software/stk/release/${P}.tar.gz"
 LICENSE="as-is"
 SLOT="0"
 KEYWORDS="~x86 ~amd64"
-IUSE="alsa jack midi oss"
+IUSE="alsa examples jack midi oss"
 
 RDEPEND="alsa? ( media-libs/alsa-lib )
-		jack? ( media-sound/jack-audio-connection-kit )
-		midi? ( media-libs/alsa-lib )"
+	jack? ( media-sound/jack-audio-connection-kit )
+	midi? ( media-libs/alsa-lib )"
 DEPEND="${RDEPEND}
-		dev-lang/perl"
+	dev-lang/perl"
+
+pkg_setup() {
+	if use midi && ! built_with_use media-libs/alsa-lib midi; then
+		eerror "You need to compile media-libs/alsa-lib with USE=\"midi\""
+		eerror "in order to get midi suppport in STK"
+	fi
+}
 
 src_unpack() {
 	unpack ${A}
@@ -37,43 +44,43 @@ src_compile() {
 		RAWWAVE_PATH=/usr/share/stk/rawwaves/ \
 		|| die "configure failed!"
 
-	# build the whole bunch
+	# compile libstk
 	cd src
 	emake || die "make in src failed!"
 
-	cd ../projects
-	for i in demo effects examples ragamatic; do
-		cd $i
-		emake || die
-		rm -f Release/*.o
-		cd ..
-	done
-
-	# reconfigure to correct paths, hell knows
-	cd "${S}"
-	econf \
-		`use_with alsa` \
-		`use_with jack` \
-		`use_with oss` \
-		RAWWAVE_PATH=/usr/share/stk/rawwaves \
-		INCLUDE_PATH=/usr/include/stk \
-		|| die "re-configure failed!"
-
+	# compile examples and projects
+	if use examples; then
+		cd ../projects
+		for i in demo effects examples ragamatic; do
+			einfo "Compiling $i ..."
+			cd $i
+			emake || die
+			rm -f Release/*.o
+			cd ..
+		done
+	fi
 	# fix include paths
 	find . -type f -name Makefile\* -exec \
 		perl -p -i -e "s|../../src|/usr/share/stk/src|g" {} \;
 }
 
 src_install() {
-	dodoc README STK_TODO.txt
+	dodoc README
+	# install the lib
+	dolib src/libstk.so*
+	# install headers
 	insinto /usr/include/stk
 	doins include/*.h include/*.msg include/*.tbl
-	insinto /usr/share/stk/src
-	doins src/*.cpp
+	# install rawwaves
 	insinto /usr/share/stk/rawwaves
 	doins rawwaves/*.raw
-	dolib src/libstk.*
-	# this one actually sucks, but..
-	mv projects "${D}"/usr/share/stk
+	if use examples; then
+		# this one actually sucks, but..
+		for i in *.bat *.cpp *.h Makefile*; do
+			find projects/ -name $i -exec rm -f {} \;
+		done
+		dodir /usr/share/"${PN}"
+		doins -r projects
+	fi
 }
